desktop: add parameters to specify window size
This adds two optional parameters, --width and --height, which
allow the user to specify a desired physical window width and
height to be used. If only one of the two parameters is present,
the other will be deduced from it and the movie's aspect ratio.

If neither are present, the window's size is set to the movie's
logical size, in accordance with the behaviour before this change.